refactor: Model에서 View로 전달되는 실행 결과 타입을 변경
